Added Switch and Scale Alliance Color to SmartDashboard/Shuffleboard
DriverStation.getGameSpecificMessage() returns a string that gives the plate randomization. Using that, I created booleans that would display the color of the plates on Smart Dashboard/Shuffleboard. Also, Shuffleboard is a better Smart Dashboard that uses the exact same methods that Smart Dashboard uses, so I think we should use it.