Post Rocket City Code + New Controller Code
Changed how autonomous worked at Rocket city so that if we are not in
the middle, and the switch is on the other side of the field, we will
just pass the auto line. DON'T CHANGE MIDDLE AUTO!!!! IT WORKS PERFECTLY
SO DON'T CHANGE IT. Added code so that the controller Adam created
works. Unsure about a good vault angle so just defaulted at 45 degrees.